+++ CHANGES.rst
@@ -16,6 +16,7 @@ Changelog
 - added `BytesSupporter` mixin
 - added the `shell-exec` reader for executing arbitrary external commands
 - added the `get-metadata` filter to extract field values from the meta-data
+- added the `sleep` filter for waiting specified number of seconds before forwarding data
 
 
 0.0.1 (2025-10-31)